# apis/sentinel/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
...
        "title": "Sentinel API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/resources": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all the resources stored",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resources"
                ],
                "summary": "Get all the resources",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Desired tenant - environment",
                        "name": "x-sentinel-tenant",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Claimant requesting the operation",
                        "name": "claimant",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/resource.Output"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/views.ErrView"
                        }
                    }
                }
            }
        },
        "/v1/resources/": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add a new resource to existing resources",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resources"
                ],
                "summary": "Create a new Resource",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Desired tenant - environment",
                        "name": "x-sentinel-tenant",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Claimant requesting the operation",
                        "name": "claimant",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Resource to be created",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/resource.Input"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/resource.OutputDetails"
                        }
                    },
                    "500": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/views.ErrView"
                        }
                    }
                }
            }
        },
        "/v1/resources/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a resource by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resources"
                ],
                "summary": "Get resource by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Desired tenant - environment",
                        "name": "x-sentinel-tenant",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Claimant requesting the operation",
                        "name": "claimant",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/resource.OutputDetails"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/views.ErrView"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/views.ErrView"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a resource by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resources"
                ],
                "summary": "Delete a resource by its ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Desired environment",
                        "name": "x-sentinel-tenant",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Claimant requesting the operation",
                        "name": "claimant",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/views.ErrView"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update resource name, sourceID, parent, etc",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resources"
                ],
                "summary": "Update a resource by it's ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Desired environment",
                        "name": "x-sentinel-tenant",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Claimant requesting the operation",
                        "name": "claimant",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Resource to be created",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/resource.Input"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/resource.Output"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/views.ErrView"
                        }
                    }
                }
            }
        },
        "/v1/resources/{id}/contexts": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all context and details that are granted to this context",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resources"
                ],
                "summary": "Get all contexts granted to this resource",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Resource ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Claimant requesting the operation",
                        "name": "claimant",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/context.Output"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/views.ErrView"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/views.ErrView"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Grant a new context to an existing principal resources",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Resources"
                ],
                "summary": "Associate a new context with an existing resource",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Desired tenant - environment",
                        "name": "x-sentinel-tenant",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Principal Resource ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "context to be created",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/context.Input"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/context.OutputDetails"
                        }
                    },
                    "500": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/views.ErrView"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "context.Attributes": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "context.Details": {
            "type": "object",
            "required": [
                "attributes",
                "type"
            ],
            "properties": {
                "attributes": {
                    "type": "object",
                    "$ref": "#/definitions/context.Attributes"
                },
                "id": {
                    "type": "string"
                },
                "relationships": {
                    "type": "object",
                    "$ref": "#/definitions/context.Relationships"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "context"
                    ]
                }
            }
        },
        "context.Input": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/context.InputDetails"
                }
            }
        },
        "context.InputDetails": {
            "type": "object",
            "required": [
                "attributes",
                "type"
            ],
            "properties": {
                "attributes": {
                    "type": "object",
                    "$ref": "#/definitions/context.Attributes"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "context"
                    ]
                }
            }
        },
        "context.Output": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/context.Details"
                    }
                }
            }
        },
        "context.OutputDetails": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/context.Details"
                }
            }
        },
        "context.Relationship": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/context.Resource"
                    }
                }
            }
        },
        "context.Relationships": {
            "type": "object",
            "required": [
                "principals",
                "target_resources"
            ],
            "properties": {
                "principals": {
                    "type": "object",
                    "$ref": "#/definitions/context.Relationship"
                },
                "target_resources": {
                    "type": "object",
                    "$ref": "#/definitions/context.Relationship"
                }
            }
        },
        "context.Resource": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "resource"
                    ]
                }
            }
        },
...
        "resource.Attributes": {
            "type": "object",
            "required": [
                "source_id"
            ],
            "properties": {
                "context_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "source_id": {
                    "type": "string"
                }
            }
        },
        "resource.Contexts": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/resource.Data"
                    }
                }
            }
        },
        "resource.Data": {
            "type": "object",
            "required": [
                "id",
                "type"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "resource"
                    ]
                }
            }
        },
        "resource.Details": {
            "type": "object",
            "properties": {
                "attributes": {
                    "type": "object",
                    "$ref": "#/definitions/resource.Attributes"
                },
                "id": {
                    "type": "string"
                },
                "relationships": {
                    "type": "object",
                    "$ref": "#/definitions/resource.Relationships"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "resource"
                    ]
                }
            }
        },
        "resource.Input": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/resource.InputDetails"
                }
            }
        },
        "resource.InputDetails": {
            "type": "object",
            "required": [
                "attributes",
                "type"
            ],
            "properties": {
                "attributes": {
                    "type": "object",
                    "$ref": "#/definitions/resource.Attributes"
                },
                "relationships": {
                    "type": "object",
                    "$ref": "#/definitions/resource.InputRelationships"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "resource"
                    ]
                }
            }
        },
        "resource.InputRelationships": {
            "type": "object",
            "required": [
                "parent"
            ],
            "properties": {
                "parent": {
                    "type": "object",
                    "$ref": "#/definitions/resource.Parent"
                }
            }
        },
        "resource.Output": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/resource.Details"
                    }
                }
            }
        },
        "resource.OutputDetails": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/resource.Details"
                }
            }
        },
        "resource.Parent": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/resource.Data"
                }
            }
        },
        "resource.Relationships": {
            "type": "object",
            "properties": {
                "contexts": {
                    "type": "object",
                    "$ref": "#/definitions/resource.Contexts"
                },
                "parent": {
                    "type": "object",
                    "$ref": "#/definitions/resource.Parent"
                }
            }
        },
        "views.ErrView": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "source": {
                    "type": "object",
                    "$ref": "#/definitions/views.Source"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "views.Source": {
            "type": "object",
            "properties": {
                "parameter": {
                    "type": "string"
                },
                "pointer": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}